Rewrite API's & writen documentation
