+++ .gitignore
@@ -1,6 +1,3 @@
 .gitigrone
-ai.php
-index.html
-OLDindex.php
 sql.txt
 .vscode/
